% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscore_year.R
\name{hscore_year}
\alias{hscore_year}
\title{A high point function}
\usage{
hscore_year(year)
}
\arguments{
\item{love}{}
}
\description{
This function returns the name of the player who scored the most points in a given year
}
\examples{
hscore_year()
}
\keyword{nba}
