% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_player.R
\name{old_player}
\alias{old_player}
\title{A old Function}
\usage{
old_player(year)
}
\arguments{
\item{love}{}
}
\description{
This function returns the name of the oldest player in a given year
}
\examples{
old_player()
}
\keyword{nba}
